regroup trick controllers in same controller
